Order confirmation process

+++ BrenMkOnline/app/src/main/java/com/brenhr/mkonline/model/Cart.kt
@@ -5,9 +5,9 @@ class Cart {
     var status: String
     var items: MutableList<ItemCart>?
     var quantity: Long
-    var total: Double
+    var total: Long
 
-    constructor(id: String,status: String, quantity: Long, total: Double) {
+    constructor(id: String,status: String, quantity: Long, total: Long) {
         this.id = id
         this.status = status
         this.items = null
