+++ BrenMkOnline/app/src/main/java/com/brenhr/mkonline/model/Order.kt
@@ -5,9 +5,9 @@ class Order {
     var status: String
     var items: MutableList<Item>?
     var quantity: Long
-    var total: Double
+    var total: Long
 
-    constructor(id: String, status: String, items: MutableList<Item>, quantity: Long, total: Double) {
+    constructor(id: String, status: String, items: MutableList<Item>, quantity: Long, total: Long) {
         this.id = id
         this.status = status
         this.items = items
@@ -15,7 +15,7 @@ class Order {
         this.total = total
     }
 
-    constructor(id: String, status: String, quantity: Long, total: Double) {
+    constructor(id: String, status: String, quantity: Long, total: Long) {
         this.id = id
         this.status = status
         this.items = null
@@ -23,7 +23,7 @@ class Order {
         this.total = total
     }
 
-    constructor(status: String, quantity: Long, total: Double) {
+    constructor(status: String, quantity: Long, total: Long) {
         this.id = null
         this.status = status
         this.items = null
